test(RHTAPWATCH-1193): template#executeTemplate
And also factor it out to its own file.
