+++ internal/template/resources.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"regexp"
 	"strings"
-	"text/template"
 
 	projctlv1beta1 "github.com/konflux-ci/project-controller/api/v1beta1"
 	"github.com/konflux-ci/project-controller/internal/ownership"
@@ -279,24 +278,4 @@ func getVarValues(
 		}
 	}
 	return
-}
-
-var nameFieldInvalidCharPattern = regexp.MustCompile("[^a-z0-9]")
-var templateFuncs = template.FuncMap{
-	"hyphenize": func(str string) string {
-		return nameFieldInvalidCharPattern.ReplaceAllString(str, "-")
-	},
-}
-
-// Execute the template given as a string and return the result as a string
-func executeTemplate(templateStr string, values map[string]string) (string, error) {
-	theTemplate, err := template.New("").Funcs(templateFuncs).Parse(templateStr)
-	if err != nil {
-		return "", err
-	}
-	var valueBuf strings.Builder
-	if err := theTemplate.Execute(&valueBuf, values); err != nil {
-		return "", err
-	}
-	return valueBuf.String(), nil
 }
